refactor: exception package 구조 변경
