+++ src/main/java/kitchenpos/table/domain/OrderTables.java
@@ -6,7 +6,7 @@ import java.util.List;
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.OneToMany;
-import kitchenpos.handler.exception.NotChangeStatusException;
+import kitchenpos.exception.NotChangeStatusException;
 import org.springframework.util.CollectionUtils;
 
 @Embeddable
