+++ src/main/java/kitchenpos/table/domain/TableGroup.java
@@ -8,7 +8,7 @@ import javax.persistence.EntityListeners;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.handler.exception.NotCreateTableGroupException;
+import kitchenpos.exception.NotCreateTableGroupException;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
