+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -3,8 +3,8 @@ package kitchenpos.menu.application;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.handler.exception.NotCreateMenuException;
-import kitchenpos.handler.exception.NotFoundEntityException;
+import kitchenpos.exception.NotCreateMenuException;
+import kitchenpos.exception.NotFoundEntityException;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuGroupRepository;
